scripts/db-sync/requirements.txt
psycopg[binary]==3.1.18
gspread==5.12.4
google-auth==2.23.4
pandas==2.1.4
python-dotenv==1.0.0
tenacity==8.2.3
